refactor: DriverHelper 구현
